+++ hw12_13_14_15_calendar/internal/app/app.go
@@ -2,6 +2,9 @@ package app
 
 import (
 	"context"
+
+	"github.com/ar2r/go-otus/hw12_13_14_15_calendar/internal/storage"
+	"github.com/google/uuid"
 )
 
 type App struct {
@@ -17,7 +20,10 @@ type Logger interface {
 }
 
 type Storage interface {
-	CreateEvent(ctx context.Context, id, title string) error
+	Add(ctx context.Context, event storage.Event) (*storage.Event, error)
+	Update(ctx context.Context, event storage.Event) (*storage.Event, error)
+	Delete(ctx context.Context, id uuid.UUID) error
+	List(ctx context.Context) ([]storage.Event, error)
 }
 
 func New(logger Logger, storage Storage) *App {
@@ -28,5 +34,6 @@ func New(logger Logger, storage Storage) *App {
 }
 
 func (a *App) CreateEvent(ctx context.Context, id, title string) error {
-	return a.Storage.CreateEvent(ctx, id, title)
+	//return a.Storage.CreateEvent(ctx, id, title)
+	return storage.ErrNotImplemented
 }
