+++ hw12_13_14_15_calendar/internal/config/app.go
@@ -1,0 +1,7 @@
+package config
+
+type AppConf struct {
+	Env     string `toml:"env"`
+	Debug   bool   `toml:"debug"`
+	Storage string `toml:"storage"`
+}
