+++ hw12_13_14_15_calendar/internal/config/database.go
@@ -1,0 +1,10 @@
+package config
+
+type DatabaseConf struct {
+	Host     string `toml:"host"`
+	Port     int    `toml:"port"`
+	Database string `toml:"database"`
+	Username string `toml:"username"`
+	Password string `toml:"password"`
+	Schema   string `toml:"schema"`
+}
