+++ hw12_13_14_15_calendar/internal/config/logger.go
@@ -1,0 +1,7 @@
+package config
+
+type LoggerConf struct {
+	Level    string `toml:"level"`
+	Channel  string `toml:"channel"`
+	Filename string `toml:"filename"`
+}
