+++ hw12_13_14_15_calendar/internal/config/server.go
@@ -1,0 +1,6 @@
+package config
+
+type ServerConf struct {
+	Host string `toml:"host"`
+	Port int    `toml:"port"`
+}
