+++ hw12_13_14_15_calendar/internal/server/http/server.go
@@ -13,7 +13,7 @@ type Logger interface { // TODO
 type Application interface { // TODO
 }
 
-func NewServer(logger Logger, app Application) *Server {
+func NewServer(app Application) *Server {
 	return &Server{}
 }
