+++ hw12_13_14_15_calendar/internal/storage/error.go
@@ -3,6 +3,7 @@ package storage
 import "errors"
 
 var (
-	ErrNotFound = errors.New("event not found")
-	ErrDateBusy = errors.New("intersecting events")
+	ErrNotFound       = errors.New("event not found")
+	ErrDateBusy       = errors.New("intersecting events")
+	ErrNotImplemented = errors.New("method not implemented")
 )
